requirements.txt
pytest==8.0.0
pyyaml==6.0.1
mcp>=1.2.0